IE-23 + oktagon: refactored code
